lazy-nutri-coach/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.2.1
python-multipart==0.0.20
sqlalchemy==2.0.29
alembic==1.13.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
httpx==0.27.2
Pillow>=11.3.0
openai==1.37.0
chromadb==0.5.3
numpy==1.26.4
pandas==2.2.2
